Fixes for BMM and BMS comms and reporting on ECU of all battery information, can also send a packet to instruct the ECU to debug different information

+++ BMM.X/SlaveCommunications.h
@@ -26,7 +26,7 @@ extern "C" {
 
     extern void updateSlaveCommunications();
 
-    extern volatile unsigned int slaveTime;
+    extern volatile unsigned long int slaveTime;
     extern void sendSlavePacket(int slave);
     extern int BVolts[NUMSLAVES][BATTPERSLAVE];
     extern int BTemps[NUMSLAVES][TEMPPERSLAVE];
